+++ pinger/utils.go
@@ -47,8 +47,8 @@ func RunPing(cfg *Config, host string) (*Result, error) {
 	}, nil
 }
 
-func convertRttsAsFloat(rtts []time.Duration) []float64 {
-	rttsFloat := make([]float64, 0, len(rtts))
+func convertRttsAsFloat(rtts []time.Duration) []float32 {
+	rttsFloat := make([]float32, 0, len(rtts))
 	for _, rtt := range rtts {
 		rttsFloat = append(rttsFloat, rtt.Seconds()*1000)
 	}
